Tests for the BighamNormalizer class
